// src/components/crud/lib/interface.ts
export type Row = { [key: string]: any };
export type MessageType =
  | ''
  | 'success'
  | 'warning'
  | 'info'
  | 'error'
  | undefined;

export interface crudStatusType {
  add: number;
  edit: number;
  cu: number;
  title: string;
}

export interface CrudOptionsType {
  crudTag?: string;
  crudPrimarykey?: string | undefined;
  crudTitle?: string;
  isPageQuery?: boolean;
  crudDelKeys?: Array<string>;
  defaultForm?: Row;
  defaultQuery?: Row;
  crudMethod?: any;
  unfixedSum?: any; // 合计不需要保存小鼠配置
  excludeSumColumn?: any; // 不需要计算合计列 key 集合 二选一
  includeSumColumn?: any; // 需要计算合计列 key 集合 二选一
  crudOptShow?: CrudOptShowType;
  setupQuery?: boolean;
  crudTableParams?: string;
  crudFormRefsKey?: string;
  crudCreateQuery?: boolean;
  downloadUrl?: string;
}
// ...
export interface CrudResDataType {
  [key: string]: any;
}

// crud方法类型
export interface CrudMethodType {
  // eslint-disable-next-line @typescript-eslint/ban-types
  list: Function;
  // eslint-disable-next-line @typescript-eslint/ban-types
  add: Function;
  // eslint-disable-next-line @typescript-eslint/ban-types
  edit: Function;
  // eslint-disable-next-line @typescript-eslint/ban-types
  del: Function;
  [key: string]: any;
}

// 按钮显示控制标识类型
/**
 * @param query 是否显示查询区域
 */
export interface CrudOptShowType {
  operate?: boolean;
  query?: boolean;
  add?: boolean;
  edit?: boolean;
  pullOrder?: boolean;
  del?: boolean;
  download?: boolean;
  import?: boolean;
  reset?: boolean;
  autoRefresh?: boolean;
  moreSet?: boolean;
  moreDel?: boolean;
  more?: boolean;
  fullscreen?: boolean;
  formSize?: string;
  dragAble?: boolean;
  crudOptShow?: boolean;
  hideFooter?: boolean;
}

export interface STATUSType {
  NORMAL: number;
  PREPARED: number;
  PROCESSING: number;
}

export interface CrudNOTIFICATION_TYPE {
  SUCCESS: MessageType;
  WARNING: MessageType;
  INFO: MessageType;
  ERROR: MessageType;
}

export interface CrudmsgType {
  submit: string;
  add: string;
  edit: string;
  del: string;
  recall: string;
  top: string;
}

export interface crudPageType {
  page: number;
  limit: number;
  count: number;
}